+++ src/index.ts
@@ -1,3 +1,16 @@
-import { startUdpSocket } from "./udpSocket";
+import dgram from "dgram";
+import { onMessage } from "./lib/handler";
+import { createOnListening } from "./lib/createOnListening";
 
-startUdpSocket();
+const server = dgram.createSocket("udp4");
+
+const PORT = 33333;
+const HOST = "0.0.0.0";
+
+server.on("listening", createOnListening(server));
+
+server.on("message", onMessage);
+
+export function startUdpSocket() {
+  server.bind(PORT, HOST);
+}
